+++ src/main/java/derek/command/taskcommand/DeadlineCommand.java
@@ -1,7 +1,8 @@
-package derek.command;
+package derek.command.taskcommand;
 
 import derek.Storage;
 import derek.Ui;
+import derek.command.taskcommand.TaskCommand;
 import derek.exception.IncorrectCommandException;
 import derek.task.Task;
